requirements.txt
transformers
datasets
streamlit
scikit-learn
newspaper3k
PyMuPDF
pytest
torch
fastapi
uvicorn
flask
torch
pandas
numpy
tqdm
sentencepiece
protobuf
requests
lxml
lxml-html-clean